# core/docs/swagger.yaml
definitions:
...
  define.GetID:
    properties:
      id:
        type: string
    type: object
...
  define.Task:
    properties:
      alarm_userids:
        description: 报警用户 多个用户
        items:
          type: string
        type: array
      alatm_status:
        description: 'alarm when task run success or fail or all all:-2 failed: -1
          success: 1'
        type: integer
      child_runparallel:
        description: 是否以并行运行子任务 0否 1是
        type: integer
      child_taskids:
        description: 子任务 运行结束后运行子任务
        items:
          type: string
        type: array
      create_by:
        description: 创建人
        type: string
      create_byuid:
        description: 创建人ID
        type: string
      create_time:
        description: 创建时间
        type: string
      cronexpr:
        description: 执行任务表达式
        type: string
      expect_code:
        description: expect task return code. if not set 0 or 200
        type: integer
      expect_content:
        description: expect task return content. if not set do not check
        type: string
      host_group:
        description: 执行计划
        type: string
      host_groupid:
        description: 主机组ID
        type: string
      id:
        type: string
      name:
        type: string
      parent_runparallel:
        description: 是否以并行运行父任务 0否 1是
        type: integer
      parent_taskids:
        description: 父任务 运行任务前先运行父任务 以父或子任务运行时 任务不会执行自已的父子任务，防止循环依赖
        items:
          type: string
        type: array
      remark:
        description: 备注
        type: string
      route_policy:
        description: how to select a run worker from hostgroup
        type: integer
      run:
        description: TODO 改为 0 后只是在调度循环中不再检测下次运行时间
        type: integer
      task_data:
        description: 任务数据
        type: object
      task_type:
        description: 任务类型
        type: integer
      timeout:
        description: 任务超时时间 (s)
        type: integer
      update_time:
        description: 最后一次更新时间
        type: string
    required:
    - alarm_userids
    - alatm_status
    - cronexpr
    - host_groupid
    - route_policy
    type: object
  resp.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
info:
  contact:
    email: labulakalia@gmail.com
    name: labulaka521
    url: http://www.swagger.io/support
  description: Crocodile Swaager JSON API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/labulaka521/crocodile
  title: Crocidle API
  version: "1.0"
paths:
...
  /api/v1/task:
    delete:
      parameters:
      - description: delete task
        in: body
        name: Task
        required: true
        schema:
          $ref: '#/definitions/define.GetID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: delete task
      tags:
      - Task
    get:
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: get tasks
      tags:
      - Task
    post:
      parameters:
      - description: create task
        in: body
        name: Task
        required: true
        schema:
          $ref: '#/definitions/define.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: create new task
      tags:
      - Task
    put:
      parameters:
      - description: change task
        in: body
        name: Task
        required: true
        schema:
          $ref: '#/definitions/define.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: change task
      tags:
      - Task
  /api/v1/task/info:
    get:
      parameters:
      - description: id
        in: query
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: get tasks
      tags:
      - Task
  /api/v1/task/kill:
    put:
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: kill running task
      tags:
      - Task
  /api/v1/task/log:
    get:
      parameters:
      - description: ID
        in: query
        name: id
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: get tasks
      tags:
      - Task
  /api/v1/task/run:
    put:
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: get tasks
      tags:
      - Task
  /api/v1/task/running:
    get:
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: get tasks
      tags:
      - Task
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"